requirements.txt
coverage>=5.3,<6.0
dj-database-url>=0.5.0
Django>=3.1.0,<4.0.0
gunicorn>=20.0.0,<21.0.0
psycopg2-binary>=2.8.0,<3.0.0
selenium>=3.141.0,<4.0.0
whitenoise>=5.2.0,<6.0.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
requests
